Changed storage and load function
Realized that "symmetric" behaviors were unnecessarily complex.  The
information on whether a load increments or decrements a storage is
inherent on if they are local or adjacent to the cell with the storage.
This means that a "negative" state is not necessary in the adjacent
boundary and does not need the resources.